Extract ArgsProcessor as separate module
